requirements.txt
fastapi
uvicorn[standard]
tensorflow
h5py
pillow
numpy
